@@ -29,14 +29,14 @@ namespace Aerospike.Test
 		public void Operate()
 		{
 			// Write initial record.
-			Key key = new Key(args.ns, args.set, "opkey");
-			Bin bin1 = new Bin("optintbin", 7);
-			Bin bin2 = new Bin("optstringbin", "string value");
+			Key key = new(SuiteHelpers.ns, SuiteHelpers.set, "opkey");
+			Bin bin1 = new("optintbin", 7);
+			Bin bin2 = new("optstringbin", "string value");
 			client.Put(null, key, bin1, bin2);
 
 			// Add integer, write new string and read record.
-			Bin bin3 = new Bin(bin1.name, 4);
-			Bin bin4 = new Bin(bin2.name, "new string");
+			Bin bin3 = new(bin1.name, 4);
+			Bin bin4 = new(bin2.name, "new string");
 			Record record = client.Operate(null, key, Operation.Add(bin3), Operation.Put(bin4), Operation.Get());
 			AssertBinEqual(key, record, bin3.name, 11);
 			AssertBinEqual(key, record, bin4);
@@ -46,8 +46,8 @@ namespace Aerospike.Test
 		public void OperateDelete()
 		{
 			// Write initial record.
-			Key key = new Key(args.ns, args.set, "opkey");
-			Bin bin1 = new Bin("optintbin1", 1);
+			Key key = new(SuiteHelpers.ns, SuiteHelpers.set, "opkey");
+			Bin bin1 = new("optintbin1", 1);
 
 			client.Put(null, key, bin1);
 
@@ -62,7 +62,7 @@ namespace Aerospike.Test
 			Assert.IsFalse(client.Exists(null, key));
 
 			// Rewrite record.
-			Bin bin2 = new Bin("optintbin2", 2);
+			Bin bin2 = new("optintbin2", 2);
 
 			client.Put(null, key, bin1, bin2);
 
@@ -79,7 +79,7 @@ namespace Aerospike.Test
 			record = client.Get(null, key);
 
 			AssertBinEqual(key, record, bin2.name, 2);
-			Assert.IsTrue(record.bins.Count == 1);
+			Assert.AreEqual(1, record.bins.Count);
 		}
 	}
 }
