@@ -30,13 +30,15 @@ namespace Aerospike.Test
 		[ClassInitialize()]
 		public static void Prepare(TestContext testContext)
 		{
-			Policy policy = new Policy();
-			policy.totalTimeout = 0; // Do not timeout on index create.
+			Policy policy = new()
+			{
+				totalTimeout = 0 // Do not timeout on index create.
+			};
 
 			try
 			{
 				IndexTask task = client.CreateIndex(
-					policy, args.ns, args.set, indexName, binName,
+					policy, SuiteHelpers.ns, SuiteHelpers.set, indexName, binName,
 					IndexType.NUMERIC, IndexCollectionType.DEFAULT,
 					CTX.ListRank(-1)
 					);
@@ -52,24 +54,19 @@ namespace Aerospike.Test
 
 			for (int i = 1; i <= size; i++)
 			{
-				Key key = new Key(args.ns, args.set, i);
+				Key key = new(SuiteHelpers.ns, SuiteHelpers.set, i);
 
-				List<int> list = new List<int>(5);
-				list.Add(i);
-				list.Add(i + 1);
-				list.Add(i + 2);
-				list.Add(i + 3);
-				list.Add(i + 4);
+				List<int> list = [i, i + 1, i + 2, i + 3, i + 4];
 
-				Bin bin = new Bin(binName, list);
+				Bin bin = new(binName, list);
 				client.Put(null, key, bin);
 			}
 		}
 
-		[ClassCleanup()]
+		[ClassCleanup(ClassCleanupBehavior.EndOfClass)]
 		public static void Destroy()
 		{
-			client.DropIndex(null, args.ns, args.set, indexName);
+			client.DropIndex(null, SuiteHelpers.ns, SuiteHelpers.set, indexName);
 		}
 
 		[TestMethod]
@@ -78,9 +75,9 @@ namespace Aerospike.Test
 			int begin = 14;
 			int end = 18;
 
-			Statement stmt = new Statement();
-			stmt.SetNamespace(args.ns);
-			stmt.SetSetName(args.set);
+			Statement stmt = new();
+			stmt.SetNamespace(SuiteHelpers.ns);
+			stmt.SetSetName(SuiteHelpers.set);
 			stmt.SetBinNames(binName);
 			stmt.SetFilter(Filter.Range(binName, begin, end, CTX.ListRank(-1)));
 
