@@ -31,49 +31,53 @@ namespace Aerospike.Test
 		[ClassInitialize()]
 		public static void Prepare(TestContext testContext)
 		{
-			WritePolicy policy = new WritePolicy();
+			WritePolicy policy = new();
 
 			// Write records in set p1.
 			for (int i = 1; i <= 5; i++)
 			{
 				policy.expiration = i * 60;
 
-				Key key = new Key(args.ns, set1, i);
+				Key key = new(SuiteHelpers.ns, set1, i);
 				client.Put(policy, key, new Bin(binA, i));
 			}
 
 			// Write records in set p2.
 			for (int i = 20; i <= 22; i++)
 			{
-				Key key = new Key(args.ns, set2, i);
+				Key key = new(SuiteHelpers.ns, set2, i);
 				client.Put(null, key, new Bin(binA, i), new Bin(binB, (double)i));
 			}
 
 			// Write records in set p3 with send key.
-			policy = new WritePolicy();
-			policy.sendKey = true;
+			policy = new WritePolicy
+			{
+				sendKey = true
+			};
 
 			for (int i = 31; i <= 40; i++)
 			{
-				Key intKey = new Key(args.ns, set3, i);
+				Key intKey = new(SuiteHelpers.ns, set3, i);
 				client.Put(policy, intKey, new Bin(binA, i));
 
-				Key strKey = new Key(args.ns, set3, "key-p3-" + i);
+				Key strKey = new(SuiteHelpers.ns, set3, "key-p3-" + i);
 				client.Put(policy, strKey, new Bin(binA, i));
 			}
 
 			// Write one record in set p3 with send key not set.
-			client.Put(null, new Key(args.ns, set3, 25), new Bin(binA, 25));
+			client.Put(null, new Key(SuiteHelpers.ns, set3, 25), new Bin(binA, 25));
 		}
 
 		[TestMethod]
 		public void QuerySetName()
 		{
-			Statement stmt = new Statement();
-			stmt.SetNamespace(args.ns);
+			Statement stmt = new();
+			stmt.SetNamespace(SuiteHelpers.ns);
 
-			QueryPolicy policy = new QueryPolicy();
-			policy.filterExp = Exp.Build(Exp.EQ(Exp.SetName(), Exp.Val(set2)));
+			QueryPolicy policy = new()
+			{
+				filterExp = Exp.Build(Exp.EQ(Exp.SetName(), Exp.Val(set2)))
+			};
 
 			RecordSet rs = client.Query(policy, stmt);
 
@@ -96,12 +100,14 @@ namespace Aerospike.Test
 		[TestMethod]
 		public void QueryDouble()
 		{
-			Statement stmt = new Statement();
-			stmt.SetNamespace(args.ns);
+			Statement stmt = new();
+			stmt.SetNamespace(SuiteHelpers.ns);
 			stmt.SetSetName(set2);
 
-			QueryPolicy policy = new QueryPolicy();
-			policy.filterExp = Exp.Build(Exp.GT(Exp.FloatBin(binB), Exp.Val(21.5)));
+			QueryPolicy policy = new()
+			{
+				filterExp = Exp.Build(Exp.GT(Exp.FloatBin(binB), Exp.Val(21.5)))
+			};
 
 			RecordSet rs = client.Query(policy, stmt);
 
@@ -124,12 +130,14 @@ namespace Aerospike.Test
 		[TestMethod]
 		public void QueryKeyString()
 		{
-			Statement stmt = new Statement();
-			stmt.SetNamespace(args.ns);
+			Statement stmt = new();
+			stmt.SetNamespace(SuiteHelpers.ns);
 			stmt.SetSetName(set3);
 
-			QueryPolicy policy = new QueryPolicy();
-			policy.filterExp = Exp.Build(Exp.RegexCompare("^key-.*-35$", 0, Exp.Key(Exp.Type.STRING)));
+			QueryPolicy policy = new()
+			{
+				filterExp = Exp.Build(Exp.RegexCompare("^key-.*-35$", 0, Exp.Key(Exp.Type.STRING)))
+			};
 
 			RecordSet rs = client.Query(policy, stmt);
 
@@ -152,12 +160,14 @@ namespace Aerospike.Test
 		[TestMethod]
 		public void QueryKeyInt()
 		{
-			Statement stmt = new Statement();
-			stmt.SetNamespace(args.ns);
+			Statement stmt = new();
+			stmt.SetNamespace(SuiteHelpers.ns);
 			stmt.SetSetName(set3);
 
-			QueryPolicy policy = new QueryPolicy();
-			policy.filterExp = Exp.Build(Exp.LT(Exp.Key(Exp.Type.INT), Exp.Val(35)));
+			QueryPolicy policy = new()
+			{
+				filterExp = Exp.Build(Exp.LT(Exp.Key(Exp.Type.INT), Exp.Val(35)))
+			};
 
 			RecordSet rs = client.Query(policy, stmt);
 
@@ -181,12 +191,14 @@ namespace Aerospike.Test
 		[TestMethod]
 		public void QueryKeyExists()
 		{
-			Statement stmt = new Statement();
-			stmt.SetNamespace(args.ns);
+			Statement stmt = new();
+			stmt.SetNamespace(SuiteHelpers.ns);
 			stmt.SetSetName(set3);
 
-			QueryPolicy policy = new QueryPolicy();
-			policy.filterExp = Exp.Build(Exp.KeyExists());
+			QueryPolicy policy = new()
+			{
+				filterExp = Exp.Build(Exp.KeyExists())
+			};
 
 			RecordSet rs = client.Query(policy, stmt);
 
@@ -209,18 +221,20 @@ namespace Aerospike.Test
 		[TestMethod]
 		public void QueryVoidTime()
 		{
-			Statement stmt = new Statement();
-			stmt.SetNamespace(args.ns);
+			Statement stmt = new();
+			stmt.SetNamespace(SuiteHelpers.ns);
 			stmt.SetSetName(set1);
 
 			DateTime now = DateTime.UtcNow;
 			DateTime end = now.Add(TimeSpan.FromMinutes(2));
 
-			QueryPolicy policy = new QueryPolicy();
-			policy.filterExp = Exp.Build(
+			QueryPolicy policy = new()
+			{
+				filterExp = Exp.Build(
 				Exp.And(
 					Exp.GE(Exp.VoidTime(), Exp.Val(now)),
-					Exp.LT(Exp.VoidTime(), Exp.Val(end))));
+					Exp.LT(Exp.VoidTime(), Exp.Val(end))))
+			};
 
 			RecordSet rs = client.Query(policy, stmt);
 
@@ -243,12 +257,14 @@ namespace Aerospike.Test
 		[TestMethod]
 		public void QueryTTL()
 		{
-			Statement stmt = new Statement();
-			stmt.SetNamespace(args.ns);
+			Statement stmt = new();
+			stmt.SetNamespace(SuiteHelpers.ns);
 			stmt.SetSetName(set1);
 
-			QueryPolicy policy = new QueryPolicy();
-			policy.filterExp = Exp.Build(Exp.LE(Exp.TTL(), Exp.Val(60)));
+			QueryPolicy policy = new()
+			{
+				filterExp = Exp.Build(Exp.LE(Exp.TTL(), Exp.Val(60)))
+			};
 
 			RecordSet rs = client.Query(policy, stmt);
 
