@@ -25,7 +25,7 @@ namespace Aerospike.Test
 	{
 		private const string indexName = "aggindex";
 		private const string keyPrefix = "aggkey";
-		private static readonly string binName = args.GetBinName("aggbin");
+		private static readonly string binName = Suite.GetBinName("aggbin");
 		private const int size = 10;
 
 		[ClassInitialize()]
@@ -35,12 +35,14 @@ namespace Aerospike.Test
 			RegisterTask task = client.Register(null, assembly, "Aerospike.Test.LuaResources.sum_example.lua", "sum_example.lua", Language.LUA);
 			task.Wait();
 
-			Policy policy = new Policy();
-			policy.totalTimeout = 0; // Do not timeout on index create.
+			Policy policy = new()
+			{
+				totalTimeout = 0 // Do not timeout on index create.
+			};
 
 			try
 			{
-				IndexTask itask = client.CreateIndex(policy, args.ns, args.set, indexName, binName, IndexType.NUMERIC);
+				IndexTask itask = client.CreateIndex(policy, SuiteHelpers.ns, SuiteHelpers.set, indexName, binName, IndexType.NUMERIC);
 				itask.Wait();
 			}
 			catch (AerospikeException ae)
@@ -53,16 +55,16 @@ namespace Aerospike.Test
 
 			for (int i = 1; i <= size; i++)
 			{
-				Key key = new Key(args.ns, args.set, keyPrefix + i);
-				Bin bin = new Bin(binName, i);
+				Key key = new(SuiteHelpers.ns, SuiteHelpers.set, keyPrefix + i);
+				Bin bin = new(binName, i);
 				client.Put(null, key, bin);
 			}
 		}
 
-		[ClassCleanup()]
+		[ClassCleanup(ClassCleanupBehavior.EndOfClass)]
 		public static void Destroy()
 		{
-			client.DropIndex(null, args.ns, args.set, indexName);
+			client.DropIndex(null, SuiteHelpers.ns, SuiteHelpers.set, indexName);
 		}
 
 		[TestMethod]
@@ -71,9 +73,9 @@ namespace Aerospike.Test
 			int begin = 4;
 			int end = 7;
 
-			Statement stmt = new Statement();
-			stmt.SetNamespace(args.ns);
-			stmt.SetSetName(args.set);
+			Statement stmt = new();
+			stmt.SetNamespace(SuiteHelpers.ns);
+			stmt.SetSetName(SuiteHelpers.set);
 			stmt.SetBinNames(binName);
 			stmt.SetFilter(Filter.Range(binName, begin, end));
 			stmt.SetAggregateFunction(Assembly.GetExecutingAssembly(), "Aerospike.Test.LuaResources.sum_example.lua", "sum_example", "sum_single_bin", Value.Get(binName));
@@ -112,16 +114,16 @@ namespace Aerospike.Test
 		[TestMethod]
 		public void QuerySetNotFound()
 		{
-			Statement stmt = new Statement()
+			Statement stmt = new()
 			{
-				Namespace = args.ns,
+				Namespace = SuiteHelpers.ns,
 				SetName = "notfound",
-				BinNames = new string[] { binName },
+				BinNames = [binName],
 				Filter = Filter.Range(binName, 4, 7)
 			};
 			stmt.SetAggregateFunction(Assembly.GetExecutingAssembly(), "Aerospike.Test.LuaResources.sum_example.lua", "sum_example", "sum_single_bin", Value.Get(binName));
 
-			QueryPolicy qp = new QueryPolicy()
+			QueryPolicy qp = new()
 			{
 				socketTimeout = 5000
 			};
