@@ -1,5 +1,5 @@
 ﻿/* 
- * Copyright 2012-2018 Aerospike, Inc.
+ * Copyright 2012-2025 Aerospike, Inc.
  *
  * Portions may be licensed to Aerospike, Inc. under one or more contributor
  * license agreements.
@@ -14,25 +14,170 @@
  * License for the specific language governing permissions and limitations under
  * the License.
  */
-using System;
-using Microsoft.VisualStudio.TestTools.UnitTesting;
+using Aerospike.Client;
 
 namespace Aerospike.Test
 {
 	[TestClass]
-	public class Suite
+	public static class Suite
 	{
 		[AssemblyInitialize()]
 		public static void AssemblyInit(TestContext context)
 		{
-			Args args = Args.Instance;
-			args.Connect();
+			SuiteHelpers.port = int.Parse(context.Properties["Port"].ToString());
+			Console.WriteLine($"Port {SuiteHelpers.port}");
+			Log.Info($"Port {SuiteHelpers.port}");
+			SuiteHelpers.clusterName = context.Properties["ClusterName"].ToString();
+			SuiteHelpers.user = context.Properties["User"].ToString();
+			SuiteHelpers.password = context.Properties["Password"].ToString();
+			SuiteHelpers.timeout = int.Parse(context.Properties["Timeout"].ToString());
+			SuiteHelpers.ns = context.Properties["Namespace"].ToString();
+			SuiteHelpers.set = context.Properties["Set"].ToString();
+			SuiteHelpers.authMode = (AuthMode)Enum.Parse(typeof(AuthMode), context.Properties["AuthMode"].ToString(), true);
+			SuiteHelpers.useServicesAlternate = bool.Parse(context.Properties["UseServicesAlternate"].ToString());
+
+			bool tlsEnable = bool.Parse(context.Properties["TlsEnable"].ToString());
+
+			if (tlsEnable)
+			{
+				SuiteHelpers.tlsName = context.Properties["TlsName"].ToString();
+				SuiteHelpers.tlsPolicy = new TlsPolicy(
+					context.Properties["TlsProtocols"].ToString(),
+					context.Properties["TlsRevoke"].ToString(),
+					context.Properties["TlsClientCertFile"].ToString(),
+					bool.Parse(context.Properties["TlsLoginOnly"].ToString())
+					);
+			}
+
+			var hostName = context.Properties["Host"].ToString();
+			if (hostName == null || hostName == String.Empty)
+			{
+				SuiteHelpers.hosts = null;
+			}
+			else
+			{
+				SuiteHelpers.hosts = Host.ParseHosts(hostName, SuiteHelpers.tlsName, SuiteHelpers.port);
+			}
+
+			ConnectSync();
+			ConnectAsync();
 		}
 
 		[AssemblyCleanup()]
 		public static void AssemblyCleanup()
 		{
-			Args.Instance.Close();
+			Close();
+		}
+
+		private static void ConnectSync()
+		{
+			ClientPolicy policy = new()
+			{
+				clusterName = SuiteHelpers.clusterName,
+				tlsPolicy = SuiteHelpers.tlsPolicy,
+				authMode = SuiteHelpers.authMode,
+				timeout = SuiteHelpers.timeout,
+				useServicesAlternate = SuiteHelpers.useServicesAlternate
+			};
+
+			if (SuiteHelpers.user != null && SuiteHelpers.user.Length > 0)
+			{
+				policy.user = SuiteHelpers.user;
+				policy.password = SuiteHelpers.password;
+			}
+
+			SuiteHelpers.client = new AerospikeClient(policy, SuiteHelpers.hosts);
+
+			//Example of how to enable metrics
+			//client.EnableMetrics(new MetricsPolicy());
+
+			try
+			{
+				SetServerSpecific();
+			}
+			catch
+			{
+				SuiteHelpers.client.Close();
+				SuiteHelpers.client = null;
+				throw;
+			}
+		}
+
+		private static void ConnectAsync()
+		{
+			AsyncClientPolicy policy = new()
+			{
+				clusterName = SuiteHelpers.clusterName,
+				tlsPolicy = SuiteHelpers.tlsPolicy,
+				authMode = SuiteHelpers.authMode,
+				asyncMaxCommands = 300,
+				timeout = SuiteHelpers.timeout,
+				useServicesAlternate = SuiteHelpers.useServicesAlternate
+			};
+
+			if (SuiteHelpers.user != null && SuiteHelpers.user.Length > 0)
+			{
+				policy.user = SuiteHelpers.user;
+				policy.password = SuiteHelpers.password;
+			}
+
+			SuiteHelpers.asyncClient = new AsyncClient(policy, SuiteHelpers.hosts);
+
+			// Example of how to enable metrics
+			//asyncClient.EnableMetrics(new MetricsPolicy());
+		}
+
+		private static void SetServerSpecific()
+		{
+			Node node = SuiteHelpers.client.Nodes[0];
+			string namespaceFilter = "namespace/" + SuiteHelpers.ns;
+			Dictionary<string, string> map = Info.Request(null, node, "edition", namespaceFilter);
+
+			string namespaceTokens = map[namespaceFilter] ?? throw new Exception(string.Format("Failed to get namespace info: host={0} namespace={1}", node, SuiteHelpers.ns));
+			SuiteHelpers.singleBin = ParseBoolean(namespaceTokens, "single-bin");
+		}
+
+		private static bool ParseBoolean(String namespaceTokens, String name)
+		{
+			string search = name + '=';
+			int begin = namespaceTokens.IndexOf(search);
+
+			if (begin < 0)
+			{
+				return false;
+			}
+
+			begin += search.Length;
+			int end = namespaceTokens.IndexOf(';', begin);
+
+			if (end < 0)
+			{
+				end = namespaceTokens.Length;
+			}
+
+			string value = namespaceTokens[begin..end];
+			return Convert.ToBoolean(value);
+		}
+
+		public static string GetBinName(string name)
+		{
+			// Single bin servers don't need a bin name.
+			return SuiteHelpers.singleBin ? "" : name;
+		}
+
+		public static void Close()
+		{
+			if (SuiteHelpers.client != null)
+			{
+				SuiteHelpers.client.Close();
+				SuiteHelpers.client = null;
+			}
+
+			if (SuiteHelpers.asyncClient != null)
+			{
+				SuiteHelpers.asyncClient.Close();
+				SuiteHelpers.asyncClient = null;
+			}
 		}
 	}
 }
